# openapi.yaml
openapi: 3.1.0
info:
  title: PPshkin API
  version: 1.0.0
  description: HTTP API мини-приложения ППшкин в MAX. Гость ведёт дневник питания
    и получает персональные предложения заведений, заведение управляет меню,
    горящими позициями и бронями.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Сессионный токен из POST /api/v1/auth/max или тестовый токен
        демо-учётки (только при DEMO_MODE=true).
  schemas:
    GeoPointInput:
      type: object
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lon:
          type: number
          minimum: -180
          maximum: 180
      required:
        - lat
        - lon
      description: Координаты WGS 84
    ProfilePatchInput:
      type: object
      properties:
        kcalTarget:
          type: integer
          minimum: 1000
          maximum: 5000
          description: Ориентир калорийности на день, ккал
        goal:
          anyOf:
            - type: string
              enum:
                - lose
                - maintain
                - gain
            - type: "null"
          description: "Цель: lose - снизить вес, maintain - удерживать, gain - набрать,
            null - не выбрана"
        dislikedTags:
          maxItems: 20
          type: array
          items:
            type: string
            enum:
              - sweet
              - dessert
              - pastry
              - chocolate
              - fruit
              - berries
              - dairy
              - cheese
              - eggs
              - meat
              - poultry
              - fish
              - seafood
              - vegetarian
              - vegetables
              - greens
              - salad
              - soup
              - grain
              - pasta
              - rice
              - potato
              - bread
              - fried
              - grilled
              - spicy
              - fast_food
              - coffee
              - tea
              - drink
              - juice
              - high_protein
              - light
              - hearty
              - breakfast
          description: Теги блюд, которые гостю не предлагаются
        timezone:
          description: Часовой пояс IANA, например Europe/Moscow
          type: string
          minLength: 1
          maxLength: 64
      description: "Изменение профиля: передаются только меняющиеся поля, хотя бы одно"
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          description: Машиночитаемый код ошибки
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required:
              - path
              - message
            additionalProperties: false
      required:
        - type
        - title
        - status
        - code
        - detail
      additionalProperties: false
      description: Описание ошибки по RFC 9457
    GeoPoint:
      type: object
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lon:
          type: number
          minimum: -180
          maximum: 180
      required:
        - lat
        - lon
      additionalProperties: false
      description: Координаты WGS 84
    ConsentState:
      type: object
      properties:
        granted:
          type: boolean
          description: true, только если действует согласие на актуальную версию текста
        version:
          description: Версия действующего согласия, null если согласия нет. Устаревшая
            версия означает, что согласие нужно дать заново
          type:
            - string
            - "null"
        grantedAt:
          anyOf:
            - type: string
              format: date-time
              examples:
                - 2026-09-25T12:00:00.000Z
              description: Дата и время в формате ISO 8601, UTC
            - type: "null"
          description: Когда дано действующее согласие
      required:
        - granted
        - version
        - grantedAt
      additionalProperties: false
      description: Состояние согласия пользователя
    ConsentStatus:
      type: object
      properties:
        personalData:
          $ref: "#/components/schemas/ConsentState"
        personalizedOffers:
          $ref: "#/components/schemas/ConsentState"
      required:
        - personalData
        - personalizedOffers
      additionalProperties: false
      description: Состояние обоих согласий пользователя
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор пользователя в MAX
        firstName:
          type:
            - string
            - "null"
        username:
          type:
            - string
            - "null"
        timezone:
          type: string
          description: Часовой пояс IANA, по нему считаются дни дневника
        kcalTarget:
          type: integer
          description: Ориентир калорийности на день, ккал
        goal:
          anyOf:
            - type: string
              enum:
                - lose
                - maintain
                - gain
            - type: "null"
          description: "Цель: lose - снизить вес, maintain - удерживать, gain - набрать,
            null - не выбрана"
        dislikedTags:
          type: array
          items:
            type: string
            enum:
              - sweet
              - dessert
              - pastry
              - chocolate
              - fruit
              - berries
              - dairy
              - cheese
              - eggs
              - meat
              - poultry
              - fish
              - seafood
              - vegetarian
              - vegetables
              - greens
              - salad
              - soup
              - grain
              - pasta
              - rice
              - potato
              - bread
              - fried
              - grilled
              - spicy
              - fast_food
              - coffee
              - tea
              - drink
              - juice
              - high_protein
              - light
              - hearty
              - breakfast
          description: Теги блюд, которые гостю не предлагаются
        location:
          anyOf:
            - $ref: "#/components/schemas/GeoPoint"
            - type: "null"
          description: Последнее переданное местоположение с точностью около 1 км
        locationUpdatedAt:
          anyOf:
            - type: string
              format: date-time
              examples:
                - 2026-09-25T12:00:00.000Z
              description: Дата и время в формате ISO 8601, UTC
            - type: "null"
          description: Когда передано местоположение
        consents:
          $ref: "#/components/schemas/ConsentStatus"
      required:
        - id
        - firstName
        - username
        - timezone
        - kcalTarget
        - goal
        - dislikedTags
        - location
        - locationUpdatedAt
        - consents
      additionalProperties: false
      description: Профиль пользователя и состояние его согласий
    UserLocation:
      type: object
      properties:
        location:
          $ref: "#/components/schemas/GeoPoint"
        updatedAt:
          type: string
          format: date-time
          examples:
            - 2026-09-25T12:00:00.000Z
          description: Дата и время в формате ISO 8601, UTC
      required:
        - location
        - updatedAt
      additionalProperties: false
      description: "Сохранённое местоположение: координаты округлены до 2 знаков, это
        около 1 км"
...
paths:
...
  /api/v1/me:
    get:
      operationId: getMe
      summary: Профиль текущего пользователя
      tags:
        - me
      description: "Профиль вместе с состоянием согласий: по consents мини-приложение
        решает, показывать ли экраны согласий. Коды ошибок: user_not_found
        (404): аккаунт удалён, войдите заново."
      security: &a3
        - bearerAuth: []
      responses:
        "200":
          description: Профиль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      operationId: updateMe
      summary: Изменить профиль
      tags:
        - me
      description: "Меняет только переданные поля: ориентир калорийности, цель,
        нелюбимые теги, часовой пояс. Повторы тегов удаляются. Коды ошибок:
        validation_failed (400): исправьте поля, перечисленные в errors.
        consent_required (403): нет согласия на обработку данных, покажите его
        текст из GET /api/v1/consents и отправьте PUT
        /api/v1/consents/personal_data. user_not_found (404): аккаунт удалён,
        войдите заново. invalid_timezone (422): часовой пояс не найден,
        передайте имя IANA, например Europe/Moscow."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilePatchInput"
      security: *a3
      responses:
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: Unprocessable Entity
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: deleteMyAccount
      summary: Удалить аккаунт и все данные
      tags:
        - me
      description: "Бесплатно и сразу удаляет профиль, дневник, согласия и состояние
        бота, отменяет активные брони с возвратом порций в горящие предложения и
        стирает объяснения показанных предложений. Повторный вызов тоже отвечает
        204. После удаления старый токен получает 404 user_not_found на GET
        /api/v1/me, новый вход создаёт пустой профиль без согласий. Коды ошибок:
        demo_account_protected (403): демо-учётка общая для всех проверяющих, её
        удалить нельзя."
      security: *a3
      responses:
        "204":
          description: Аккаунт удалён
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/me/location:
    put:
      operationId: setMyLocation
      summary: Сохранить местоположение
      tags:
        - me
      description: "Координаты округляются до 2 знаков (около 1 км) и сохраняются
        только в таком виде, по ним подбираются заведения рядом. Коды ошибок:
        validation_failed (400): исправьте поля, перечисленные в errors.
        consent_required (403): нет согласия на обработку данных, покажите его
        текст из GET /api/v1/consents и отправьте PUT
        /api/v1/consents/personal_data. user_not_found (404): аккаунт удалён,
        войдите заново."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeoPointInput"
      security: *a3
      responses:
        "200":
          description: Сохранённое местоположение
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLocation"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: clearMyLocation
      summary: Удалить местоположение
      tags:
        - me
      description: Работает без согласия на обработку данных. Повторный вызов тоже
        отвечает 204.
      security: *a3
      responses:
        "204":
          description: Местоположение удалено
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
servers:
  - url: /
    description: Текущий хост
...